Add accountID parameter to transaction methods

+++ internal/entities/quote_entity.go
@@ -9,7 +9,7 @@ import (
 
 type Quote struct {
 	ID           string    `gorm:"primaryKey;type:varchar(50)"` // เช่น q_1234...
-	AccountID    int64     `gorm:"index"`
+	AccountID    uint      `gorm:"index"`
 	Type         string    `gorm:"type:varchar(20)"`                   // เช่น "TOPUP_CRYPTO"
 	PromptPayID  *string   `gorm:"column:promptpay_id;default:null"`   // เช่น "0812345678"
 	THBAmount    int64     `gorm:"column:thb_amount"`                  // เก็บเป็นสตางค์ เช่น 100000 (1,000 บาท)
